fitXiao and Feing configurable tc and flexible wMax
both fitXiao and fitFeingoldModBound now have two features:
tc can be fixed to an arbitrary value
wMax is now flexible: it can be used for normalized data as well as
absolute diameters/radii in nm. the initial guess (and upper and lower
bounds) are determined by the first value for w.