getFWXM speed up
Replaced superfluous repetitive call to interp1 by a single splinefit.
Results in a speed increase of almost an order of magnitude, depending
on the data.
Note that it now takes the FWXM of the spline instead of the
interpolated raw data, which is better in the case of the SIManalysis,
where the data has a weird stepwise behaviour. (implying that the
interpolation is not always meaningful).