add fitBilinear
function to fit a bilinear model to input data.
Defined as a line with offset (A) and slope (B), with the slope changing (to C) at a certain point (tc).
The two lines are continuous, their value at tc is the same. (A, B and tc determine the offset of the second line with slope C).